requirements.txt
agno==2.5.10
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
certifi==2026.2.25
click==8.3.1
distro==1.9.0
docstring_parser==0.17.0
gitdb==4.0.12
GitPython==3.1.46
groq==1.1.1
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
packaging==26.0
pydantic==2.12.5
pydantic-settings==2.13.1
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
shellingham==1.5.4
smmap==5.0.3
sniffio==1.3.1
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
